+++ test/lib/perfect_shape/test_affine_transform.rb
@@ -317,6 +317,9 @@ describe PerfectShape do
       xt = 0 # point x coordinate translation (m13)
       yt = 0 # point y coordinate translation (m23)
       affine_transform = PerfectShape::AffineTransform.new(xxp: xxp, xyp: xyp, yxp: yxp, yyp: yyp, xt: xt, yt: yt)
+      
+      assert affine_transform.invertible?
+      
       affine_transform.invert!
       
       _(affine_transform.xxp).must_equal 2
@@ -325,6 +328,19 @@ describe PerfectShape do
       _(affine_transform.yyp).must_equal 2
       _(affine_transform.xt).must_equal 0
       _(affine_transform.yt).must_equal 0
+    end
+    
+    it 'cannot invert affine transform with non-invertible matrix ' do
+      xxp = 2 # point x coordinate x product (m11)
+      xyp = 4 # point x coordinate y product (m12)
+      yxp = 2 # point y coordinate x product (m21)
+      yyp = 4 # point y coordinate y product (m22)
+      xt = 0 # point x coordinate translation (m13)
+      yt = 0 # point y coordinate translation (m23)
+      affine_transform = PerfectShape::AffineTransform.new(xxp: xxp, xyp: xyp, yxp: yxp, yyp: yyp, xt: xt, yt: yt)
+      
+      refute affine_transform.invertible?
+      _(proc {affine_transform.invert!}).must_raise StandardError
     end
   end
 end